Fix SA1402 violation in DrawingsLabDataSource.cs
